1) Add a "default" type to print a message if nothing previously matched at that level or since the last default at that level.  This is useful for setting up switch-like statements.  It can also be used to do if/else constructions without a redundant second test.
2) Fix the "x" special case test so that one can test for that string
with "=x".

3) Allow "search" to search the entire buffer if the "/N" search count
is missing.

4) Make "regex" work!  It now starts its search at the specified
offset and takes an (optional) "/N" line count to specify the search
range; otherwise it searches to the end of the file.  The match is now
grabbed correctly for format strings and the offset set to the end of
the match.

5) Add a "/s" flag to "regex" and "search" to set the offset to the
start of the match.  By default the offset is set to the end of the
match, as it is with other tests.  This is mostly useful for "regex".

6) Make "search", "string" and "pstring" use the same file_strncmp()
routine so that they support the same flags; "bestring16" and
"lestring16" call the same routine, but with flags = 0.  Also add a
"/C" flag (in analogy to "/c") to ignore the case on uppercase
(lowercase) characters in the test string.

7) Strict adherence to C style string escapes.  A warnings are
printed when compiling.  Note: previously "\a" was incorrectly
translated to 'a' instead of an <alert> (i.e., BELL, typically 0x07).

8) Make this compile with "-Wall -Wextra" and all the warning flags
used with WARNS=4 in the NetBSD source.  Also make it pass lint.

9) Many "cleanups" and hopefully not too many new bugs!