Change test harness so that it uses the main magic file, rather than per-test magic.
It would be easy to reintroduce per-test magic, but neither the one
(!) existing test, nor the one being added, needs it.

Add a new test for a bug recently fixed in EPUB magic. Let's have more
tests!

Making this change uncovered a failure: the magic for GEDCOM files has
been changed in the main magic files, and doesn't work any more. This
is bug #315. The magic should work as-is, but since the real fix is
non-obvious, change it back.

Update the gedcom.result, since the output has also changed in the
main magic files.