From Christoph Biedl: Debian bug #745546 et al: At the moment, the magic for "Macromedia Flash data" and "Macromedia Flash Video" is just three printable characters, no wonder this results in mis-detection of text files.
"Flash Video" is easy to fix, according to Wikipedia the forth
octet must have value 1.

"Flash data" is more work. The length information at offset 4 must be
non-zero, and the version number at offset 3 must be sound. According
to the pdf document, the latest version is 19, I found a few examples
in the lower twenties, so 32 should be a sane limit for the time
being. There are however quite a few files with version 255, and since
the virus scanners created an alert for all of them, they are now
reported as "suspicious".