From Christoph Biedl:
The bad news first: Detection of WordPerfect/Corel is quite broken due
to forgotten endianess clarification, and has probably been ever since.
On LE archs, nothing but "(Corel/WP)" is printed.

Time for a clarification and cleanup. Basically, there's a product
(Offset 8) and file type (Offset 9) short, and little information about
their meanings. So I took the old logik and placed it in logical
groups, with some questions about PlanPerfect raised. Perhaps there was
an off-by-one when creating the data? Also, the "loadable file" and
"graphic image" fit into the scheme, no need for extra rules.

There were a few hints in the wp2x sources, and the LibreOffice has
a few sample files in the sources. Not all of them are detected, these
have a completely different header (very old format?).