switch http:// to https:// where possible
This was done by scanning each http:// domain with gnutls-cli.  if it
responded successfully, then i replace all the http:// links with
https://

There are still some http:// links remaining, of course, but many
fewer.  It's possible that some of the remaining links (and some of
the transformed links) are dead URLs.  This process didn't try to
detect that particular problem.

domains=$(git grep 'http://.' | sed -e 's_^.*http://__' -e 's_[^-a-zA-Z0-9.].*$__' | sort | uniq)
for x in $domains ; do
  if echo | gnutls-cli $x >/dev/null 2>/dev/null; then
    pattern="s_http://${x}_https://${x}_"
    echo "swapping $x" "$pattern"
    sed -i  "$pattern" $(git grep -l "http://$x")
  else
    echo skipping $x
  fi &
done
wait