Refine ARM uncompressed Kernel pattern endianness.
For a while now its been possible to configure the kernel to have little
endian magic values, but still be a big endian kernel. This is to allow
different endianness between the bootloader and the kernel.

Parse the new magic flag has been added to help differentiate the two.
(Jonathan Duncan)